Move project validation tests to test file
